server/requirements.txt
fastapi[standard]
numpy
pandas
scikit-learn
scipy
httpx
python-dotenv
